+++ data-provider/src/main/scala/velocorner/api/brand/ProductDetails.scala
@@ -15,5 +15,5 @@ case class ProductDetails(
     price: Money,
     imageUrl: String,
     productUrl: String,
-    reviewStars: Int
+    reviewStars: Double
 )
